Refactor cli/index to use run-series
Read --scheduled option to not reload scheduled events unless activated from cli
