+++ src/helpers/index.js
@@ -2,11 +2,13 @@ let env = require('./env')
 let { getPorts, checkPort } = require('./ports')
 let maybeHydrate = require('./maybe-hydrate')
 let readArc = require('./read-arc')
+let readOptions = require('./read-options')
 
 module.exports = {
   env,
   getPorts,
   checkPort,
   maybeHydrate,
-  readArc
+  readArc,
+  readOptions
 }
